refactor: move async generator utility to lib directory
