+++ src/actions/search.ts
@@ -1,5 +1,5 @@
 import { fetchAllRevisions } from '../services/MediaWikiAPIs';
-import { genFindMap } from '../utils/async-generator';
+import { genFindMap } from '../lib/async-generator';
 import { type SearchState } from '../state';
 import { WIKI_SITES } from '../wiki';
 
@@ -136,7 +136,7 @@ if (import.meta.vitest) {
 	const { describe, it, expect, vi, beforeEach } = import.meta.vitest;
 
 	vi.mock(import('../services/MediaWikiAPIs'), { spy: true });
-	vi.mock(import('../utils/async-generator'), { spy: true });
+	vi.mock(import('../lib/async-generator'), { spy: true });
 
 	async function* createAsyncGenerator<T>(
 		items: ReadonlyArray<T>
